+++ vms/secp256k1fx/keychain.go
@@ -124,7 +124,7 @@ func (kc *Keychain) Spend(out verify.Verifiable, time uint64) (verify.Verifiable
 // Match attempts to match a list of addresses up to the provided threshold.
 // Assumes [time] is a Unix time > 0.
 func (kc *Keychain) Match(owners *OutputOwners, time uint64) ([]uint32, []*secp256k1.PrivateKey, bool) {
-	if uint64(time) < owners.Locktime {
+	if time < owners.Locktime {
 		return nil, nil, false
 	}
 	sigs := make([]uint32, 0, owners.Threshold)
